refactor: move 'ShoppingCart' component in components-directory
